Separate code to show search dialog in utils method
