+++ src/main/java/org/quantumbadger/redreader/activities/CommentListingActivity.java
@@ -17,20 +17,18 @@
 
 package org.quantumbadger.redreader.activities;
 
-import android.app.AlertDialog;
-import android.content.DialogInterface;
 import android.content.Intent;
 import android.net.Uri;
 import android.os.Bundle;
+import android.support.annotation.Nullable;
 import android.util.Log;
 import android.view.Menu;
 import android.view.MenuItem;
-import android.view.WindowManager;
-import android.widget.EditText;
 
 import org.quantumbadger.redreader.R;
 import org.quantumbadger.redreader.account.RedditAccountChangeListener;
 import org.quantumbadger.redreader.account.RedditAccountManager;
+import org.quantumbadger.redreader.common.DialogUtils;
 import org.quantumbadger.redreader.common.General;
 import org.quantumbadger.redreader.common.LinkHandler;
 import org.quantumbadger.redreader.common.PrefsUtility;
@@ -176,29 +174,13 @@ public class CommentListingActivity extends RefreshableActivity
 
 	@Override
 	public void onSearchComments() {
-
-		final AlertDialog.Builder alertBuilder = new AlertDialog.Builder(this);
-		final EditText editText = (EditText) getLayoutInflater().inflate(R.layout.dialog_editbox, null);
-
-		alertBuilder.setView(editText);
-		alertBuilder.setTitle(R.string.action_search);
-
-		alertBuilder.setPositiveButton(R.string.action_search, new DialogInterface.OnClickListener() {
+		DialogUtils.showSearchDialog(this, new DialogUtils.OnSearchListener() {
 			@Override
-			public void onClick(DialogInterface dialog, int which) {
-
-				final String query = editText.getText().toString().toLowerCase().trim();
+			public void onSearch(@Nullable String query) {
 				controller.setSearchString(query);
-
-				requestRefresh(RefreshableFragment.COMMENTS, false);
+				requestRefresh(RefreshableActivity.RefreshableFragment.COMMENTS, false);
 			}
 		});
-
-		alertBuilder.setNegativeButton(R.string.dialog_cancel, null);
-
-		final AlertDialog alertDialog = alertBuilder.create();
-		alertDialog.getWindow().setSoftInputMode(WindowManager.LayoutParams.SOFT_INPUT_STATE_VISIBLE);
-		alertDialog.show();
 	}
 
 	@Override
